invoice added successfully

+++ src/main/java/com/krishnkant/inventorybackendflow/cart/dto/CartResponseDTO.java
@@ -1,10 +1,11 @@
 package com.krishnkant.inventorybackendflow.cart.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartResponseDTO(
         Long cartId,
         String status,
         List<CartItemResponseDTO> items,
-        Double totalAmount
+        BigDecimal totalAmount
 ) {}
