+++ src/main/java/com/krishnkant/inventorybackendflow/order/dto/OrderResponseDTO.java
@@ -1,13 +1,14 @@
 package com.krishnkant.inventorybackendflow.order.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderResponseDTO(
         Long orderId,
         String orderReference,
-        Double totalAmount,
-        Double discountAmount,
-        Double finalAmount,
+        BigDecimal totalAmount,
+        BigDecimal discountAmount,
+        BigDecimal finalAmount,
         String status,
         List<OrderItemDTO> items
 ) {}
