+++ src/main/java/com/krishnkant/inventorybackendflow/product/dto/ProductResponseDTO.java
@@ -1,9 +1,11 @@
 package com.krishnkant.inventorybackendflow.product.dto;
 
+import java.math.BigDecimal;
+
 public record ProductResponseDTO(
         Long id,
         String name,
-        Double price,
+        BigDecimal price,
         Integer stock,
         Boolean active
 ) {}
